pyproject.toml
[project]
name = "uvproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "cryptography>=45.0.5",
    "easyocr>=1.7.2",
    "fastapi>=0.116.1",
    "fuzzywuzzy>=0.18.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-experimental>=0.3.4",
    "langchain-groq>=0.3.6",
    "langgraph>=0.6.1",
    "matplotlib>=3.10.5",
    "openai>=1.97.1",
    "opencv-contrib-python>=4.12.0.88",
    "opencv-python>=4.12.0.88",
    "opencv-python-headless>=4.12.0.88",
    "passlib>=1.7.4",
    "pillow>=11.3.0",
    "psycopg2>=2.9.10",
    "pydantic[email]>=2.11.7",
    "pymysql>=1.1.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "rapidfuzz>=3.14.0",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.42",
    "streamlit>=1.47.1",
    "tabulate>=0.9.0",
    "tesseract>=0.1.3",
    "uvicorn>=0.35.0",
]
